requirements.txt
accelerate==0.28.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dateparser==1.2.0
emojis==0.7.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
htmldate==1.8.0
huggingface-hub==0.22.1
idna==3.6
igraph==0.11.4
importlib_resources==6.3.1
it-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/it_core_news_lg-3.7.0/it_core_news_lg-3.7.0-py3-none-any.whl#sha256=f48bd152621c872c1f177dbe21929fbb28751e73eb3c61714cf6344c6d582bbf
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.42.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
peft==0.10.0
pillow==10.2.0
preshed==3.0.9
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.2
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.19.0
shap==0.45.1
six==1.16.0
slicer==0.0.8
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.36.3
sympy==1.12
textstat==0.7.3
texttable==1.7.0
thinc==8.2.3
threadpoolctl==3.3.0
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.29.0
wasabi==1.1.2
weasel==0.3.4
whois==1.20240129.2
wsproto==1.2.0
zipp==3.18.1
